requirements.txt
Django==4.2.11
firebase-admin==6.2.0
python-dotenv==1.0.0
gunicorn==22.0.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
discord.py==2.3.2